Move Ecto repo to event store application
