+++ apps/absence/lib/absence/application.ex
@@ -4,9 +4,7 @@ defmodule Absence.Application do
   use Application
 
   def start(_type, _args) do
-    children = [
-      Absence.Repo
-    ]
+    children = []
 
     opts = [strategy: :one_for_one, name: Absence.Supervisor]
     Supervisor.start_link(children, opts)
