--- apps/absence/lib/absence/repo.ex
@@ -1,5 +1,0 @@
-defmodule Absence.Repo do
-  use Ecto.Repo,
-    otp_app: :absence,
-    adapter: Ecto.Adapters.Postgres
-end
