+++ apps/absence_web/config/config.exs
@@ -6,7 +6,7 @@ use Mix.Config
 
 # General application configuration
 config :absence_web,
-  ecto_repos: [Absence.Repo],
+  ecto_repos: [EventSourcing.EventStore.Repo],
   generators: [context_app: :absence, binary_id: true]
 
 # Configures the endpoint
