+++ apps/absence_web/test/support/channel_case.ex
@@ -26,10 +26,10 @@ defmodule AbsenceWeb.ChannelCase do
   end
 
   setup tags do
-    :ok = Ecto.Adapters.SQL.Sandbox.checkout(Absence.Repo)
+    :ok = Ecto.Adapters.SQL.Sandbox.checkout(EventSourcing.EventStore.Repo)
 
     unless tags[:async] do
-      Ecto.Adapters.SQL.Sandbox.mode(Absence.Repo, {:shared, self()})
+      Ecto.Adapters.SQL.Sandbox.mode(EventSourcing.EventStore.Repo, {:shared, self()})
     end
 
     :ok
