+++ apps/event_sourcing/lib/event_sourcing/application.ex
@@ -7,6 +7,7 @@ defmodule EventSourcing.Application do
     children = [
       EventSourcing.Aggregate.AggregateSupervisor,
       EventSourcing.EventHandler,
+      EventSourcing.EventStore.Repo,
       {Registry, keys: :unique, name: EventSourcing.AggregateRegistry}
     ]
