+++ apps/event_sourcing/lib/event_sourcing/event_store/repo.ex
@@ -1,0 +1,5 @@
+defmodule EventSourcing.EventStore.Repo do
+  use Ecto.Repo,
+    otp_app: :event_sourcing,
+    adapter: Ecto.Adapters.Postgres
+end
