+++ apps/event_sourcing/priv/repo/seeds.exs
@@ -5,7 +5,7 @@
 # Inside the script, you can read and write to any of your
 # repositories directly:
 #
-#     Absence.Repo.insert!(%Absence.SomeSchema{})
+#     EventSourcing.EventStore.Repo.insert!(%Absence.SomeSchema{})
 #
 # We recommend using the bang functions (`insert!`, `update!`
 # and so on) as they will fail if something goes wrong.
