+++ apps/event_sourcing/test/support/data_case.ex
@@ -1,4 +1,4 @@
-defmodule Absence.DataCase do
+defmodule EventSourcing.DataCase do
   @moduledoc """
   This module defines the setup for tests requiring
   access to the application's data layer.
@@ -16,20 +16,20 @@ defmodule Absence.DataCase do
 
   using do
     quote do
-      alias Absence.Repo
+      alias EventSourcing.EventStore.Repo
 
       import Ecto
       import Ecto.Changeset
       import Ecto.Query
-      import Absence.DataCase
+      import EventSourcing.DataCase
     end
   end
 
   setup tags do
-    :ok = Ecto.Adapters.SQL.Sandbox.checkout(Absence.Repo)
+    :ok = Ecto.Adapters.SQL.Sandbox.checkout(EventSourcing.EventStore.Repo)
 
     unless tags[:async] do
-      Ecto.Adapters.SQL.Sandbox.mode(Absence.Repo, {:shared, self()})
+      Ecto.Adapters.SQL.Sandbox.mode(EventSourcing.EventStore.Repo, {:shared, self()})
     end
 
     :ok
