+++ apps/event_sourcing/test/test_helper.exs
@@ -1,3 +1,4 @@
 {:ok, _} = EventSourcing.EventStore.AgentEventStore.start_link([])
 
 ExUnit.start()
+Ecto.Adapters.SQL.Sandbox.mode(EventSourcing.EventStore.Repo, :manual)
